AP2234 move delegated function methods to a mixin
